push new cutout/threshold
push new cutout/threshold, and other updates on webpages
