+++ DEMO_channelf.f90
