+++ DEMO_getCutoutc.c
