+++ DEMO_getCutoutf.f90
