+++ DEMO_mixingf.f90
