+++ TurbulenceService.h
@@ -1,6 +1,7 @@
+// Reminder: Modify typemap.dat to customize the header file generated by wsdl2h
 /* TurbulenceService.h
-   Generated by wsdl2h 2.8.16 from http://test.turbulence.pha.jhu.edu/service/turbulence.asmx?WSDL and typemap.dat
-   2019-07-08 22:09:49 GMT
+   Generated by wsdl2h 2.8.16 from http://turbulence.pha.jhu.edu/service/turbulence.asmx?WSDL and typemap.dat
+   2019-09-16 16:23:17 GMT
 
    DO NOT INCLUDE THIS FILE DIRECTLY INTO YOUR PROJECT BUILDS
    USE THE soapcpp2-GENERATED SOURCE CODE FILES FOR YOUR PROJECT BUILDS
@@ -1932,7 +1933,7 @@ struct _turb1__GetAnyCutoutWebResponse
 
 //gsoap turb1 service name:	TurbulenceServiceSoap 
 //gsoap turb1 service type:	TurbulenceServiceSoap 
-//gsoap turb1 service port:	http://test.turbulence.pha.jhu.edu/service/turbulence.asmx 
+//gsoap turb1 service port:	http://turbulence.pha.jhu.edu/service/turbulence.asmx 
 //gsoap turb1 service namespace:	http://turbulence.pha.jhu.edu/ 
 //gsoap turb1 service transport:	http://schemas.xmlsoap.org/soap/http 
 
@@ -2139,7 +2140,7 @@ struct _turb1__GetAnyCutoutWebResponse
 
 @section TurbulenceServiceSoap_ports Endpoints of Binding  "TurbulenceServiceSoap"
 
-  - http://test.turbulence.pha.jhu.edu/service/turbulence.asmx
+  - http://turbulence.pha.jhu.edu/service/turbulence.asmx
 
 Note: use wsdl2h option -Nname to change the service binding prefix name
